# Avoid closures in CachedTypeInspector